{"schema":"decodex/commit/1","summary":"Split progressive search follow-up APIs into a sibling module after strict local validation.","authority":"manual"}
